requirements.txt
py-cord
langchain
langchain-community
langchain-google-genai
langchain-core
langchain-chroma
chromadb
python-dotenv
